+++ src/main/java/me/marcarrots/trivia/utils/Elapsed.java
@@ -2,7 +2,11 @@
  * Trivia by MarCarrot, 2020
  */
 
-package me.marcarrots.trivia;
+/*
+ * Trivia by MarCarrot, 2020
+ */
+
+package me.marcarrots.trivia.utils;
 
 public class Elapsed {
