+++ src/main/java/me/marcarrots/trivia/utils/StringSimilarity.java
@@ -1,4 +1,4 @@
-package me.marcarrots.trivia.api;
+package me.marcarrots.trivia.utils;
 
 import org.apache.commons.lang.StringUtils;
