+++ src/main/java/me/marcarrots/trivia/utils/UpdateChecker.java
@@ -1,7 +1,7 @@
 /*
  * Trivia by MarCarrot, 2020
  */
-package me.marcarrots.trivia.api;
+package me.marcarrots.trivia.utils;
 
 import me.marcarrots.trivia.Trivia;
 import org.bukkit.Bukkit;
